requirements.txt
python-dotenv==0.10.3
Django==3.1.3
djangorestframework==3.12.2
Markdown==3.3.3
django-filter==2.4.0
drf-yasg==1.20.0
psycopg2-binary==2.8.4
Pillow==8.0.0
django-json-widget==1.0.1
gunicorn==20.0.4
django-rest-knox==4.1.0
rest-social-auth==4.2.0
celery==5.0.5
django-celery-results==2.0.0
transitions==0.8.5
django-rest-auth==0.9.5
django-allauth==0.44.0
Faker
